chore: minor changes to font
