--- src/components/theme.ts
@@ -1,14 +1,0 @@
-import { createTheme } from "@mui/material";
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#ffffff',
-    },
-    secondary: {
-      main: '#f50057',
-    },
-  }
-})
-
-export { theme };
